+++ R/condition_fun.R
@@ -54,26 +54,36 @@ check_cateCols_uniqueValues = function(dt, var_skip = NULL) {
 }
 
 # replace blank by NA
+#' @importFrom stringi stri_isempty
 rep_blank_na = function(dt) {
   dt = setDT(dt)
 
-  if (any(dt == "", na.rm = TRUE)) {
-    warning(sprintf('The blank values are replaced with NAs in the following columns:\n%s', paste(names(which(dt[,sapply(.SD, function(x) any(grepl('^\\s*$', x), na.rm = T))])), collapse = ", ")))
+  # replace black values
+  char_cols = names(which(dt[,sapply(.SD, function(x) is.character(x) | is.factor(x))]))
+  if (length(char_cols)>0) {
+    # columns have blank value
+    blank_cols = names(which(dt[, sapply(.SD, function(x) any(stri_isempty(x))), .SD = char_cols])) # grep('^\\s*$', x)
+    # repalce by NA
+    if (length(blank_cols)>0) {
+      warning(sprintf('The blank values are replaced with NAs in the following columns:\n%s', paste(blank_cols, collapse = ", ")))
 
-    dt[dt == ""] = NA
+      dt[, (blank_cols) := lapply(.SD, function(x) {
+        x[stri_isempty(x)] = NA # # grep('^\\s*$', x)
+        return(x)
+      }), .SD = blank_cols]#[dt == ""] = NA
+    }
   }
 
-  is_inf_nan = sapply(dt, function(x) any(is.infinite(x)) | any(is.nan(x)))
-  if (any(is_inf_nan)) {
-    warning(sprintf('The Infinite or NaN values are replaced with -999 in the following columns:\n%s', paste(names(which(
-      is_inf_nan
-    )), collapse = ", ")))
 
-    dt[dt == Inf | dt == -Inf] = -999
-    dt = dt[, lapply(.SD, function(x) {
-      x[is.nan(x)] = -999
+  # repalce infinite or NaN values
+  cols_inf_nan = names(which(dt[, sapply(.SD, function(x) any(is.infinite(x)) | any(is.nan(x)))]))
+  if (length(cols_inf_nan) > 0) {
+    warning(sprintf('The Infinite or NaN values are replaced with -999 in the following columns:\n%s', paste(cols_inf_nan, collapse = ", ")))
+
+    dt[, (cols_inf_nan) := lapply(.SD, function(x) {
+      x[is.nan(x) | is.infinite(x)] = -999
       return(x)
-    })]
+    }), .SD = cols_inf_nan]
   }
 
   return(dt)